Edge-Bundling-@shaestha
Creating a Edge bundling d3.js chart where it represents to overcome the problem of clustering when we visualize high categories of data. In this, every hierarchical category is adjacent to each other in a circle, and then thread-like structures work as connections. Each node has a connection to its associated category, which makes it look less hefty.